# Batch completer + additional completer test suite and benchmarks