--- a/internal/jobcompleter/main_test.go
+++ b/internal/jobcompleter/main_test.go
@@ -1,0 +1,11 @@
+package jobcompleter
+
+import (
+	"testing"
+
+	"github.com/riverqueue/river/internal/riverinternaltest"
+)
+
+func TestMain(m *testing.M) {
+	riverinternaltest.WrapTestMain(m)
+}