@brandur 2024-03-10
Actually, I ended up hacking my benchmarker into this branch. 5x+ improvement at >35k jobs/sec on my commodity MBA (compared to 6k jobs/sec previously).

Before:

``` sh
$ go run main.go bench --database-url $DATABASE_URL --duration 1m
bench: jobs worked [          0 ], inserted [      50000 ], job/sec [        0.0 ] [0s]
bench: jobs worked [       3634 ], inserted [       4000 ], job/sec [     1817.0 ] [2s]
bench: jobs worked [       4230 ], inserted [       4000 ], job/sec [     2115.0 ] [2s]
bench: jobs worked [       4238 ], inserted [       4000 ], job/sec [     2119.0 ] [2s]
bench: jobs worked [       4309 ], inserted [       6000 ], job/sec [     2154.5 ] [2s]
bench: jobs worked [       4251 ], inserted [       4000 ], job/sec [     2125.5 ] [2s]
bench: jobs worked [       4246 ], inserted [       4000 ], job/sec [     2123.0 ] [2s]
bench: jobs worked [       4202 ], inserted [       4000 ], job/sec [     2101.0 ] [2s]
bench: jobs worked [       3982 ], inserted [       4000 ], job/sec [     1991.0 ] [2s]
bench: jobs worked [       4055 ], inserted [       4000 ], job/sec [     2027.5 ] [2s]
bench: jobs worked [       4060 ], inserted [       4000 ], job/sec [     2030.0 ] [2s]
bench: jobs worked [       4100 ], inserted [       4000 ], job/sec [     2050.0 ] [2s]
bench: jobs worked [       3985 ], inserted [       4000 ], job/sec [     1992.5 ] [2s]
bench: jobs worked [       4002 ], inserted [       4000 ], job/sec [     2001.0 ] [2s]
bench: jobs worked [       4016 ], inserted [       4000 ], job/sec [     2008.0 ] [2s]
bench: jobs worked [       3834 ], inserted [       4000 ], job/sec [     1917.0 ] [2s]
bench: jobs worked [       3775 ], inserted [       4000 ], job/sec [     1887.5 ] [2s]
bench: jobs worked [       3815 ], inserted [       4000 ], job/sec [     1907.5 ] [2s]
bench: jobs worked [       3617 ], inserted [       4000 ], job/sec [     1808.5 ] [2s]
bench: jobs worked [       9917 ], inserted [      10000 ], job/sec [     4958.5 ] [2s]
bench: jobs worked [       7186 ], inserted [       6000 ], job/sec [     3593.0 ] [2s]
bench: jobs worked [      25564 ], inserted [      26000 ], job/sec [    12782.0 ] [2s]
bench: jobs worked [      24033 ], inserted [      24000 ], job/sec [    12016.5 ] [2s]
bench: jobs worked [      27612 ], inserted [      26000 ], job/sec [    13806.0 ] [2s]
bench: jobs worked [      26952 ], inserted [      28000 ], job/sec [    13476.0 ] [2s]
bench: jobs worked [      27083 ], inserted [      28000 ], job/sec [    13541.5 ] [2s]
bench: jobs worked [      27501 ], inserted [      26000 ], job/sec [    13750.5 ] [2s]
bench: jobs worked [      27636 ], inserted [      28000 ], job/sec [    13818.0 ] [2s]
bench: jobs worked [      27914 ], inserted [      28000 ], job/sec [    13957.0 ] [2s]
bench: jobs worked [      28998 ], inserted [      30000 ], job/sec [    14499.0 ] [2s]
bench: total jobs worked [     366572 ], total jobs inserted [     416000 ], overall job/sec [     6109.5 ], running 1m0.00017125s
```

After:

``` sh
$ go run main.go bench --database-url $DATABASE_URL --duration 1m
bench: jobs worked [          0 ], inserted [      50000 ], job/sec [        0.0 ] [0s]
bench: jobs worked [      50005 ], inserted [      52000 ], job/sec [    25002.5 ] [2s]
bench: jobs worked [      79998 ], inserted [      74000 ], job/sec [    39999.0 ] [2s]
bench: jobs worked [      70013 ], inserted [      76000 ], job/sec [    35006.5 ] [2s]
bench: jobs worked [      92621 ], inserted [      88000 ], job/sec [    46310.5 ] [2s]
bench: jobs worked [      87398 ], inserted [      86000 ], job/sec [    43699.0 ] [2s]
bench: jobs worked [      61233 ], inserted [      66000 ], job/sec [    30616.5 ] [2s]
bench: jobs worked [      90732 ], inserted [      76000 ], job/sec [    45366.0 ] [2s]
bench: jobs worked [      62004 ], inserted [      78000 ], job/sec [    31002.0 ] [2s]
bench: jobs worked [      93999 ], inserted [      86000 ], job/sec [    46999.5 ] [2s]
bench: jobs worked [      83998 ], inserted [      82000 ], job/sec [    41999.0 ] [2s]
bench: jobs worked [      73999 ], inserted [      78000 ], job/sec [    36999.5 ] [2s]
bench: jobs worked [      71581 ], inserted [      76000 ], job/sec [    35790.5 ] [2s]
bench: jobs worked [      76939 ], inserted [      78000 ], job/sec [    38469.5 ] [2s]
bench: jobs worked [      73105 ], inserted [      68000 ], job/sec [    36552.5 ] [2s]
bench: jobs worked [      80354 ], inserted [      86000 ], job/sec [    40177.0 ] [2s]
bench: jobs worked [      70004 ], inserted [      60000 ], job/sec [    35002.0 ] [2s]
bench: jobs worked [      70011 ], inserted [      76000 ], job/sec [    35005.5 ] [2s]
bench: jobs worked [      84107 ], inserted [      84000 ], job/sec [    42053.5 ] [2s]
bench: jobs worked [      65873 ], inserted [      70000 ], job/sec [    32936.5 ] [2s]
bench: jobs worked [      80005 ], inserted [      72000 ], job/sec [    40002.5 ] [2s]
bench: jobs worked [      70009 ], inserted [      74000 ], job/sec [    35004.5 ] [2s]
bench: jobs worked [      80015 ], inserted [      78000 ], job/sec [    40007.5 ] [2s]
bench: jobs worked [      70016 ], inserted [      76000 ], job/sec [    35008.0 ] [2s]
bench: jobs worked [      69993 ], inserted [      70000 ], job/sec [    34996.5 ] [2s]
bench: jobs worked [      80020 ], inserted [      80000 ], job/sec [    40010.0 ] [2s]
bench: jobs worked [      90008 ], inserted [      80000 ], job/sec [    45004.0 ] [2s]
bench: jobs worked [      60011 ], inserted [      70000 ], job/sec [    30005.5 ] [2s]
bench: jobs worked [      70566 ], inserted [      70000 ], job/sec [    35283.0 ] [2s]
bench: jobs worked [      79411 ], inserted [      80000 ], job/sec [    39705.5 ] [2s]
bench: total jobs worked [    2268151 ], total jobs inserted [    2314000 ], overall job/sec [    37802.2 ], running 1m0.000476583s
```