@brandur 2024-03-16
Made a couple improvements and large changes, so here's a repost of some benchmark runs. One thing I found too is that I'm able to get more consistent results when operating the benchmark in "burn down" mode where we insert all jobs before starting.

* Before: ~10k jobs/sec
* After: ~46k jobs/sec

Run before changes:

``` sh
$ go build . && ./river bench --database-url $DATABASE_URL -n 1_000_000
bench: jobs worked [          0 ], inserted [    1000000 ], job/sec [        0.0 ] [0s]
bench: jobs worked [       9087 ], inserted [          0 ], job/sec [     4543.5 ] [2s]
bench: jobs worked [      10453 ], inserted [          0 ], job/sec [     5226.5 ] [2s]
bench: jobs worked [       8310 ], inserted [          0 ], job/sec [     4155.0 ] [2s]
bench: jobs worked [      11418 ], inserted [          0 ], job/sec [     5709.0 ] [2s]
bench: jobs worked [      10875 ], inserted [          0 ], job/sec [     5437.5 ] [2s]
bench: jobs worked [      10899 ], inserted [          0 ], job/sec [     5449.5 ] [2s]
bench: jobs worked [      10768 ], inserted [          0 ], job/sec [     5384.0 ] [2s]
bench: jobs worked [      11502 ], inserted [          0 ], job/sec [     5751.0 ] [2s]
bench: jobs worked [      11247 ], inserted [          0 ], job/sec [     5623.5 ] [2s]
bench: jobs worked [      10870 ], inserted [          0 ], job/sec [     5435.0 ] [2s]
bench: jobs worked [      11146 ], inserted [          0 ], job/sec [     5573.0 ] [2s]
bench: jobs worked [      10862 ], inserted [          0 ], job/sec [     5431.0 ] [2s]
bench: jobs worked [      11006 ], inserted [          0 ], job/sec [     5503.0 ] [2s]
bench: jobs worked [      10604 ], inserted [          0 ], job/sec [     5302.0 ] [2s]
bench: jobs worked [       8877 ], inserted [          0 ], job/sec [     4438.5 ] [2s]
bench: jobs worked [      10624 ], inserted [          0 ], job/sec [     5312.0 ] [2s]
bench: jobs worked [      10843 ], inserted [          0 ], job/sec [     5421.5 ] [2s]
bench: jobs worked [      16371 ], inserted [          0 ], job/sec [     8185.5 ] [2s]
bench: jobs worked [      18023 ], inserted [          0 ], job/sec [     9011.5 ] [2s]
bench: jobs worked [      21967 ], inserted [          0 ], job/sec [    10983.5 ] [2s]
bench: jobs worked [      29516 ], inserted [          0 ], job/sec [    14758.0 ] [2s]
bench: jobs worked [      27671 ], inserted [          0 ], job/sec [    13835.5 ] [2s]
bench: jobs worked [      28273 ], inserted [          0 ], job/sec [    14136.5 ] [2s]
bench: jobs worked [      31026 ], inserted [          0 ], job/sec [    15513.0 ] [2s]
bench: jobs worked [      30648 ], inserted [          0 ], job/sec [    15324.0 ] [2s]
bench: jobs worked [      31054 ], inserted [          0 ], job/sec [    15527.0 ] [2s]
bench: jobs worked [      28779 ], inserted [          0 ], job/sec [    14389.5 ] [2s]
bench: jobs worked [      27188 ], inserted [          0 ], job/sec [    13594.0 ] [2s]
bench: jobs worked [      24629 ], inserted [          0 ], job/sec [    12314.5 ] [2s]
bench: jobs worked [      19879 ], inserted [          0 ], job/sec [     9939.5 ] [2s]
bench: jobs worked [      25347 ], inserted [          0 ], job/sec [    12673.5 ] [2s]
bench: jobs worked [      27751 ], inserted [          0 ], job/sec [    13875.5 ] [2s]
bench: jobs worked [      28440 ], inserted [          0 ], job/sec [    14220.0 ] [2s]
bench: jobs worked [      29316 ], inserted [          0 ], job/sec [    14658.0 ] [2s]
bench: jobs worked [      28189 ], inserted [          0 ], job/sec [    14094.5 ] [2s]
bench: jobs worked [      29762 ], inserted [          0 ], job/sec [    14881.0 ] [2s]
bench: jobs worked [      27336 ], inserted [          0 ], job/sec [    13668.0 ] [2s]
bench: jobs worked [      26620 ], inserted [          0 ], job/sec [    13310.0 ] [2s]
bench: jobs worked [      27051 ], inserted [          0 ], job/sec [    13525.5 ] [2s]
bench: jobs worked [      26063 ], inserted [          0 ], job/sec [    13031.5 ] [2s]
bench: jobs worked [      25188 ], inserted [          0 ], job/sec [    12594.0 ] [2s]
bench: jobs worked [      27466 ], inserted [          0 ], job/sec [    13733.0 ] [2s]
bench: jobs worked [      27301 ], inserted [          0 ], job/sec [    13650.5 ] [2s]
bench: jobs worked [      26436 ], inserted [          0 ], job/sec [    13218.0 ] [2s]
bench: jobs worked [      25720 ], inserted [          0 ], job/sec [    12860.0 ] [2s]
bench: jobs worked [      25168 ], inserted [          0 ], job/sec [    12584.0 ] [2s]
bench: jobs worked [      26160 ], inserted [          0 ], job/sec [    13080.0 ] [2s]
bench: jobs worked [      23750 ], inserted [          0 ], job/sec [    11875.0 ] [2s]
bench: jobs worked [       2519 ], inserted [          0 ], job/sec [     1259.5 ] [2s]
bench: total jobs worked [    1000000 ], total jobs inserted [    1000000 ], overall job/sec [    10204.1 ], running 1m38.00023s
```

And two runs from after the changes, showing roughly similar results:

``` sh
$ go build . && ./river bench --database-url $DATABASE_URL -n 1_000_000
bench: jobs worked [          0 ], inserted [    1000000 ], job/sec [        0.0 ] [0s]
bench: jobs worked [      88058 ], inserted [          0 ], job/sec [    44029.0 ] [2s]
bench: jobs worked [      94047 ], inserted [          0 ], job/sec [    47023.5 ] [2s]
bench: jobs worked [      90256 ], inserted [          0 ], job/sec [    45128.0 ] [2s]
bench: jobs worked [      91427 ], inserted [          0 ], job/sec [    45713.5 ] [2s]
bench: jobs worked [     100908 ], inserted [          0 ], job/sec [    50454.0 ] [2s]
bench: jobs worked [      94504 ], inserted [          0 ], job/sec [    47252.0 ] [2s]
bench: jobs worked [      81588 ], inserted [          0 ], job/sec [    40794.0 ] [2s]
bench: jobs worked [      94824 ], inserted [          0 ], job/sec [    47412.0 ] [2s]
bench: jobs worked [      87340 ], inserted [          0 ], job/sec [    43670.0 ] [2s]
bench: jobs worked [      92049 ], inserted [          0 ], job/sec [    46024.5 ] [2s]
bench: jobs worked [      84999 ], inserted [          0 ], job/sec [    42499.5 ] [2s]
bench: total jobs worked [    1000000 ], total jobs inserted [    1000000 ], overall job/sec [    45851.7 ], running 21.809460917s
```

``` sh
$ go build . && ./river bench --database-url $DATABASE_URL -n 1_000_000
bench: jobs worked [          0 ], inserted [    1000000 ], job/sec [        0.0 ] [0s]
bench: jobs worked [      82657 ], inserted [          0 ], job/sec [    41328.5 ] [2s]
bench: jobs worked [      96057 ], inserted [          0 ], job/sec [    48028.5 ] [2s]
bench: jobs worked [      89829 ], inserted [          0 ], job/sec [    44914.5 ] [2s]
bench: jobs worked [      96847 ], inserted [          0 ], job/sec [    48423.5 ] [2s]
bench: jobs worked [      96042 ], inserted [          0 ], job/sec [    48021.0 ] [2s]
bench: jobs worked [      87198 ], inserted [          0 ], job/sec [    43599.0 ] [2s]
bench: jobs worked [      96474 ], inserted [          0 ], job/sec [    48237.0 ] [2s]
bench: jobs worked [      94126 ], inserted [          0 ], job/sec [    47063.0 ] [2s]
bench: jobs worked [      85323 ], inserted [          0 ], job/sec [    42661.5 ] [2s]
bench: jobs worked [      94043 ], inserted [          0 ], job/sec [    47021.5 ] [2s]
bench: jobs worked [      81387 ], inserted [          0 ], job/sec [    40693.5 ] [2s]
bench: total jobs worked [    1000000 ], total jobs inserted [    1000000 ], overall job/sec [    45753.1 ], running 21.856442959s
```